[models] refactor: split the trajectory bridge by responsibility
The single trajectory_bridge module had grown to cover stored-trajectory
reading, forward-process noising, replayed-step dispatch, and component
reduction at once. Split it into a package with one module per stable
responsibility; the public surface is re-exported unchanged, so importers
are untouched and every function body is identical to before.
