--- src/flow_factory/models/trajectory_bridge.py
@@ -1,1214 +1,0 @@
-# Copyright 2026 Jayce-Ping
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import Any, Dict, Mapping, Optional, Tuple, Union
-
-import torch
-from diffusers.utils.torch_utils import randn_tensor
-
-from ..samples import (
-    ComponentTimes,
-    LatentState,
-    MultiModalStepOutput,
-    NoisedState,
-    ReplayStep,
-    StackedSampleBatch,
-    StructuredTrajectory,
-)
-from ..scheduler import SDESchedulerOutput
-from ..utils.base import filter_kwargs, to_broadcast_tensor
-from ..utils.noise_schedule import flow_match_sigma
-from .latent_geometry import LatentAxes
-
-_STORAGE_KEYS = {
-    "trajectory",
-    "timesteps",
-    "all_latents",
-    "latent_index_map",
-    "log_probs",
-    "log_prob_index_map",
-}
-# Trainer-owned batch fields: written by the feedback stage for the loss, never a
-# model conditioning argument. Adapters that accept ``**kwargs`` would otherwise
-# receive them, which legacy trainers never did.
-_TRAINER_METADATA_KEYS = {"advantage"}
-_BRIDGE_OWNED_BATCH_KEYS = _STORAGE_KEYS | _TRAINER_METADATA_KEYS
-_STATE_OWNED_FORWARD_KEYS = {
-    "t",
-    "t_next",
-    "latents",
-    "next_latents",
-    "compute_log_prob",
-    "return_kwargs",
-    "noise_level",
-}
-_SIGNED_INTEGER_DTYPES = {
-    torch.int8,
-    torch.int16,
-    torch.int32,
-    torch.int64,
-}
-
-
-def _require_legacy_tensors(
-    batch: StackedSampleBatch,
-    names: Tuple[str, ...],
-    *,
-    operation: str,
-    step_index: Optional[int] = None,
-) -> None:
-    missing = tuple(name for name in names if batch.get(name) is None)
-    context = f" at step_index={step_index}" if step_index is not None else ""
-    if missing:
-        raise ValueError(
-            f"expected legacy fields {names} for {operation}{context}, received keys "
-            f"{tuple(sorted(batch.keys()))} with missing values {missing}"
-        )
-    received_types = {
-        name: type(batch[name]).__name__
-        for name in names
-        if not isinstance(batch[name], torch.Tensor)
-    }
-    if received_types:
-        raise TypeError(
-            f"expected torch.Tensor legacy fields {names} for {operation}{context}, "
-            f"received types {received_types}"
-        )
-
-
-def _map_position(
-    index_map: torch.Tensor,
-    map_name: str,
-    rollout_position: int,
-    *,
-    upper_bound: int,
-) -> int:
-    if index_map.ndim != 1 or index_map.dtype not in _SIGNED_INTEGER_DTYPES:
-        raise TypeError(
-            f"expected signed integer {map_name} with shape (T,), received "
-            f"dtype {index_map.dtype} and shape {tuple(index_map.shape)}"
-        )
-    if rollout_position >= index_map.shape[0]:
-        raise ValueError(
-            f"expected rollout position below {index_map.shape[0]} for {map_name}, "
-            f"received {rollout_position} with map contents {index_map.tolist()}"
-        )
-    stored_position = int(index_map[rollout_position].item())
-    if stored_position == -1:
-        raise ValueError(
-            f"{map_name} at rollout position {rollout_position} received uncollected "
-            f"sentinel -1; map contents {index_map.tolist()}"
-        )
-    if stored_position < 0 or stored_position >= upper_bound:
-        raise ValueError(
-            f"{map_name} at rollout position {rollout_position} expected stored index in "
-            f"[0, {upper_bound - 1}], received {stored_position}; map contents "
-            f"{index_map.tolist()}"
-        )
-    return stored_position
-
-
-def _trajectory_value_at(values: torch.Tensor, position: int, *, identifier: str) -> torch.Tensor:
-    if values.ndim != 2:
-        raise ValueError(
-            f"expected batched {identifier} shape (B, T), received {tuple(values.shape)}"
-        )
-    if position >= values.shape[1]:
-        raise ValueError(
-            f"expected {identifier} position below {values.shape[1]}, received {position}"
-        )
-    return values[:, position]
-
-
-def _legacy_next_timestep(timesteps: torch.Tensor, step_index: int) -> torch.Tensor:
-    """Return the stored next timestep, or the terminal ``0`` the legacy replay used.
-
-    Adapters store one timestep per denoising step ``(B, T)`` while latents keep
-    ``T + 1`` rollout positions, so the final transition has no stored ``t_next``.
-    """
-    if timesteps.ndim != 2:
-        raise ValueError(
-            f"expected batched legacy timesteps shape (B, T), received {tuple(timesteps.shape)}"
-        )
-    if step_index + 1 >= timesteps.shape[1]:
-        return torch.tensor(0, device=timesteps.device)
-    return timesteps[:, step_index + 1]
-
-
-def _validate_structured_trajectory(
-    adapter: Any,
-    batch: StackedSampleBatch,
-    trajectory: StructuredTrajectory,
-) -> None:
-    if not isinstance(batch, StackedSampleBatch):
-        raise TypeError(
-            "expected StackedSampleBatch for structured trajectory bridge, "
-            f"received {type(batch).__name__}"
-        )
-    expected_names = adapter.trajectory_component_order
-    if trajectory.component_names != expected_names:
-        raise ValueError(
-            f"expected trajectory_component_order {expected_names}, received structured "
-            f"component order {trajectory.component_names}"
-        )
-    expected_batch_size = len(batch.samples)
-    for name in expected_names:
-        component = trajectory.components[name]
-        states_shape = tuple(component.states.shape)
-        timesteps_shape = tuple(component.timesteps.shape)
-        if (
-            component.states.ndim < 2
-            or component.timesteps.ndim != 2
-            or component.states.shape[0] != expected_batch_size
-            or component.timesteps.shape[0] != expected_batch_size
-        ):
-            raise ValueError(
-                f"expected batched StructuredTrajectory component {name!r} with states "
-                f"shape (B, stored, ...) and timesteps shape (B, T), received states "
-                f"{states_shape} and timesteps {timesteps_shape} for batch size "
-                f"{expected_batch_size}"
-            )
-        if component.sigmas is not None and component.sigmas.shape != component.timesteps.shape:
-            raise ValueError(
-                f"expected batched sigmas matching timesteps for component {name!r}, "
-                f"received sigmas {tuple(component.sigmas.shape)} and timesteps "
-                f"{timesteps_shape}"
-            )
-    if trajectory.log_probs is not None and (
-        trajectory.log_probs.ndim != 2 or trajectory.log_probs.shape[0] != expected_batch_size
-    ):
-        raise ValueError(
-            "expected batched StructuredTrajectory.log_probs shape (B, T), received "
-            f"{tuple(trajectory.log_probs.shape)} for batch size {expected_batch_size}"
-        )
-
-
-def _structured_active_masks(
-    adapter: Any,
-    trajectory: StructuredTrajectory,
-    *,
-    operation: str,
-) -> Optional[Dict[str, torch.Tensor]]:
-    """Collect every component's static active mask, or none when unmasked."""
-    expected_names = adapter.trajectory_component_order
-    present = tuple(
-        name for name in expected_names if trajectory.components[name].active_mask is not None
-    )
-    if not present:
-        return None
-    missing = tuple(name for name in expected_names if name not in present)
-    if missing:
-        raise ValueError(
-            f"expected an active_mask on all components {expected_names} or on none for "
-            f"{operation}, received masks for {present} with missing components {missing}"
-        )
-    return {name: trajectory.components[name].active_mask for name in expected_names}
-
-
-def _expand_active_mask(
-    mask: torch.Tensor,
-    values: torch.Tensor,
-    *,
-    component: str,
-    operation: str,
-) -> torch.Tensor:
-    """Broadcast one static mask onto the values it selects."""
-    if mask.ndim != values.ndim or any(
-        mask_dim not in (1, value_dim) for mask_dim, value_dim in zip(mask.shape, values.shape)
-    ):
-        raise ValueError(
-            f"expected the {operation} active mask for component {component!r} to broadcast onto "
-            f"the value shape {tuple(values.shape)}, received mask shape {tuple(mask.shape)}"
-        )
-    return mask.expand(values.shape)
-
-
-def _active_element_counts(
-    expanded_mask: torch.Tensor,
-    *,
-    component: str,
-) -> torch.Tensor:
-    """Return one positive active element count per sample."""
-    counts = expanded_mask.reshape(expanded_mask.shape[0], -1).sum(dim=1)
-    minimum = int(counts.min().item())
-    if minimum <= 0:
-        raise ValueError(
-            f"expected component {component!r} active_mask to leave a positive active element "
-            f"count per sample, received {minimum} for at least one sample with counts "
-            f"{counts.tolist()}"
-        )
-    return counts
-
-
-def _get_terminal_state(adapter: Any, batch: StackedSampleBatch) -> LatentState:
-    trajectory = batch.get("trajectory")
-    if trajectory is not None:
-        if not isinstance(trajectory, StructuredTrajectory):
-            raise TypeError(
-                "expected StructuredTrajectory or None for batch['trajectory'], "
-                f"received {type(trajectory).__name__}"
-            )
-        _validate_structured_trajectory(adapter, batch, trajectory)
-        terminal: Dict[str, torch.Tensor] = {}
-        for name in adapter.trajectory_component_order:
-            component = trajectory.components[name]
-            terminal_position = _map_position(
-                component.state_index_map,
-                f"component {name!r} state_index_map",
-                component.state_index_map.shape[0] - 1,
-                upper_bound=component.states.shape[1],
-            )
-            terminal[name] = component.states[:, terminal_position]
-        return LatentState(
-            terminal,
-            active_masks=_structured_active_masks(
-                adapter, trajectory, operation="get_terminal_state"
-            ),
-        )
-
-    _require_legacy_tensors(
-        batch,
-        ("all_latents", "latent_index_map"),
-        operation="get_terminal_state",
-    )
-    terminal_position = _map_position(
-        batch["latent_index_map"],
-        "latent_index_map",
-        batch["latent_index_map"].shape[0] - 1,
-        upper_bound=batch["all_latents"].shape[1],
-    )
-    return LatentState({"latent": batch["all_latents"][:, terminal_position]})
-
-
-def _get_replay_step(
-    adapter: Any,
-    batch: StackedSampleBatch,
-    step_index: int,
-) -> ReplayStep:
-    if not isinstance(step_index, int):
-        raise TypeError(
-            f"expected int step_index for get_replay_step, received "
-            f"{type(step_index).__name__}: {step_index!r}"
-        )
-    if step_index < 0:
-        raise ValueError(f"expected non-negative step_index, received {step_index}")
-
-    trajectory = batch.get("trajectory")
-    if trajectory is not None:
-        if not isinstance(trajectory, StructuredTrajectory):
-            raise TypeError(
-                "expected StructuredTrajectory or None for batch['trajectory'], "
-                f"received {type(trajectory).__name__} at step_index={step_index}"
-            )
-        _validate_structured_trajectory(adapter, batch, trajectory)
-        state: Dict[str, torch.Tensor] = {}
-        next_state: Dict[str, torch.Tensor] = {}
-        timestep: Dict[str, torch.Tensor] = {}
-        next_timestep: Dict[str, torch.Tensor] = {}
-        sigma: Dict[str, torch.Tensor] = {}
-        next_sigma: Dict[str, torch.Tensor] = {}
-        sigma_presence = []
-        for name in adapter.trajectory_component_order:
-            component = trajectory.components[name]
-            state_position = _map_position(
-                component.state_index_map,
-                f"component {name!r} state_index_map",
-                step_index,
-                upper_bound=component.states.shape[1],
-            )
-            next_state_position = _map_position(
-                component.state_index_map,
-                f"component {name!r} state_index_map",
-                step_index + 1,
-                upper_bound=component.states.shape[1],
-            )
-            state[name] = component.states[:, state_position]
-            next_state[name] = component.states[:, next_state_position]
-            timestep[name] = _trajectory_value_at(
-                component.timesteps,
-                step_index,
-                identifier=f"component {name!r} timesteps",
-            )
-            next_timestep[name] = _trajectory_value_at(
-                component.timesteps,
-                step_index + 1,
-                identifier=f"component {name!r} timesteps",
-            )
-            sigma_presence.append(component.sigmas is not None)
-            if component.sigmas is not None:
-                sigma[name] = _trajectory_value_at(
-                    component.sigmas,
-                    step_index,
-                    identifier=f"component {name!r} sigmas",
-                )
-                next_sigma[name] = _trajectory_value_at(
-                    component.sigmas,
-                    step_index + 1,
-                    identifier=f"component {name!r} sigmas",
-                )
-        if len(set(sigma_presence)) != 1:
-            raise ValueError(
-                "expected all structured trajectory components to provide sigmas or none, "
-                f"received {dict(zip(adapter.trajectory_component_order, sigma_presence))} "
-                f"at step_index={step_index}"
-            )
-        log_prob = None
-        component_log_probs: Optional[Dict[str, torch.Tensor]] = None
-        if trajectory.log_probs is not None:
-            if trajectory.log_prob_index_map is None:
-                log_prob_position = step_index
-            else:
-                log_prob_position = _map_position(
-                    trajectory.log_prob_index_map,
-                    "structured log_prob_index_map",
-                    step_index,
-                    upper_bound=trajectory.log_probs.shape[1],
-                )
-            log_prob = _trajectory_value_at(
-                trajectory.log_probs,
-                log_prob_position,
-                identifier="structured log_probs",
-            )
-            if trajectory.component_log_probs is not None:
-                component_log_probs = {
-                    name: _trajectory_value_at(
-                        trajectory.component_log_probs[name],
-                        log_prob_position,
-                        identifier=f"structured component {name!r} log_probs",
-                    )
-                    for name in adapter.trajectory_component_order
-                }
-        active_masks = _structured_active_masks(
-            adapter, trajectory, operation=f"get_replay_step at step_index={step_index}"
-        )
-        return ReplayStep(
-            state=LatentState(state, active_masks=active_masks),
-            next_state=LatentState(next_state, active_masks=active_masks),
-            times=ComponentTimes(
-                timestep=timestep,
-                next_timestep=next_timestep,
-                sigma=sigma if sigma_presence[0] else None,
-                next_sigma=next_sigma if sigma_presence[0] else None,
-            ),
-            log_prob=log_prob,
-            component_log_probs=component_log_probs,
-        )
-
-    _require_legacy_tensors(
-        batch,
-        ("all_latents", "latent_index_map", "timesteps"),
-        operation="get_replay_step",
-        step_index=step_index,
-    )
-    state_position = _map_position(
-        batch["latent_index_map"],
-        "latent_index_map",
-        step_index,
-        upper_bound=batch["all_latents"].shape[1],
-    )
-    next_state_position = _map_position(
-        batch["latent_index_map"],
-        "latent_index_map",
-        step_index + 1,
-        upper_bound=batch["all_latents"].shape[1],
-    )
-    log_prob = None
-    if batch.get("log_probs") is not None:
-        _require_legacy_tensors(
-            batch,
-            ("log_probs", "log_prob_index_map"),
-            operation="get_replay_step log probability",
-            step_index=step_index,
-        )
-        log_prob_position = _map_position(
-            batch["log_prob_index_map"],
-            "log_prob_index_map",
-            step_index,
-            upper_bound=batch["log_probs"].shape[1],
-        )
-        log_prob = _trajectory_value_at(
-            batch["log_probs"],
-            log_prob_position,
-            identifier="legacy log_probs",
-        )
-    return ReplayStep(
-        state=LatentState({"latent": batch["all_latents"][:, state_position]}),
-        next_state=LatentState({"latent": batch["all_latents"][:, next_state_position]}),
-        times=ComponentTimes(
-            timestep={
-                "latent": _trajectory_value_at(
-                    batch["timesteps"], step_index, identifier="legacy timesteps"
-                )
-            },
-            next_timestep={"latent": _legacy_next_timestep(batch["timesteps"], step_index)},
-        ),
-        log_prob=log_prob,
-        component_log_probs=None if log_prob is None else {"latent": log_prob},
-    )
-
-
-def _shared_callback_index_map(batch: StackedSampleBatch, field: str) -> torch.Tensor:
-    index_map = batch["callback_index_map"]
-    if index_map.ndim == 1:
-        return index_map
-    if index_map.ndim != 2:
-        raise ValueError(
-            f"expected legacy callback_index_map shape (T,) or (B, T) for callback {field!r}, "
-            f"received {tuple(index_map.shape)}"
-        )
-    if not bool(torch.equal(index_map, index_map[:1].expand_as(index_map))):
-        raise ValueError(
-            f"expected one shared legacy callback_index_map across the batch for callback "
-            f"{field!r}, received per-sample maps {index_map.tolist()}"
-        )
-    return index_map[0]
-
-
-def _get_replay_callback(
-    adapter: Any,
-    batch: StackedSampleBatch,
-    step_index: int,
-    field: str,
-) -> LatentState:
-    if not isinstance(step_index, int) or isinstance(step_index, bool):
-        raise TypeError(
-            f"expected int step_index for get_replay_callback, received "
-            f"{type(step_index).__name__}: {step_index!r}"
-        )
-    if step_index < 0:
-        raise ValueError(f"expected non-negative step_index, received {step_index}")
-    if not isinstance(field, str) or not field:
-        raise ValueError(
-            f"expected a non-empty callback field name, received {field!r} "
-            f"at step_index={step_index}"
-        )
-
-    trajectory = batch.get("trajectory")
-    if trajectory is not None:
-        if not isinstance(trajectory, StructuredTrajectory):
-            raise TypeError(
-                "expected StructuredTrajectory or None for batch['trajectory'], "
-                f"received {type(trajectory).__name__} at step_index={step_index}"
-            )
-        _validate_structured_trajectory(adapter, batch, trajectory)
-        if trajectory.callbacks is None or field not in trajectory.callbacks:
-            raise ValueError(
-                f"expected structured callback field {field!r} in stored fields "
-                f"{trajectory.callback_fields} at step_index={step_index}"
-            )
-        stored = trajectory.callbacks[field]
-        expected_names = adapter.trajectory_component_order
-        if tuple(stored) != expected_names:
-            raise ValueError(
-                f"expected callback {field!r} component order {expected_names}, received "
-                f"{tuple(stored)} at step_index={step_index}"
-            )
-        expected_batch_size = len(batch.samples)
-        components: Dict[str, torch.Tensor] = {}
-        for name in expected_names:
-            indexed = stored[name]
-            if not indexed.batched or indexed.values.shape[0] != expected_batch_size:
-                raise ValueError(
-                    f"expected batched callback {field!r} component {name!r} with batch size "
-                    f"{expected_batch_size}, received batched={indexed.batched} and values "
-                    f"shape {tuple(indexed.values.shape)}"
-                )
-            components[name] = indexed.at(
-                step_index, identifier=f"callback {field!r} component {name!r}"
-            )
-        return LatentState(
-            components,
-            active_masks=_structured_active_masks(
-                adapter,
-                trajectory,
-                operation=f"get_replay_callback {field!r} at step_index={step_index}",
-            ),
-        )
-
-    _require_legacy_tensors(
-        batch,
-        (field, "callback_index_map"),
-        operation="get_replay_callback",
-        step_index=step_index,
-    )
-    index_map = _shared_callback_index_map(batch, field)
-    stored_position = _map_position(
-        index_map,
-        f"callback {field!r} callback_index_map",
-        step_index,
-        upper_bound=batch[field].shape[1],
-    )
-    return LatentState({"latent": batch[field][:, stored_position]})
-
-
-def _get_state_active_numel(adapter: Any, state: LatentState) -> Dict[str, int]:
-    if not isinstance(state, LatentState):
-        raise TypeError(
-            f"expected LatentState for get_state_active_numel, received {type(state).__name__}"
-        )
-    expected_names = adapter.trajectory_component_order
-    if state.component_names != expected_names:
-        raise ValueError(
-            "expected state component keys/order to match trajectory_component_order "
-            f"{expected_names}, received {state.component_names}"
-        )
-    active_numel: Dict[str, int] = {}
-    for name in expected_names:
-        values = state.components[name]
-        if values.ndim < 2 or values.shape[0] <= 0:
-            raise ValueError(
-                f"expected batched state component {name!r} with shape (B, ...) and a positive "
-                f"batch size, received {tuple(values.shape)}"
-            )
-        if state.active_masks is None:
-            active_numel[name] = int(values.numel() // values.shape[0])
-            continue
-        expanded = _expand_active_mask(
-            state.active_masks[name],
-            values,
-            component=name,
-            operation="get_state_active_numel",
-        )
-        counts = _active_element_counts(expanded, component=name)
-        # The public result is one integer per component, so a per-sample count
-        # would silently pick a single sample's geometry for the whole batch.
-        if int(counts.min().item()) != int(counts.max().item()):
-            raise ValueError(
-                f"expected component {name!r} active_mask to mark a constant active element "
-                f"count per sample, received {counts.tolist()}"
-            )
-        active_numel[name] = int(counts[0].item())
-    return active_numel
-
-
-def _get_train_step_indices(adapter: Any) -> torch.Tensor:
-    group = adapter.scheduler_group
-    primary_name = group.primary_name
-    indices = group.primary.train_timesteps
-    if not isinstance(indices, torch.Tensor) or indices.ndim != 1:
-        raise TypeError(
-            f"expected 1-D torch.Tensor train step indices from scheduler component "
-            f"{primary_name!r}, received {type(indices).__name__}"
-        )
-    if indices.dtype not in _SIGNED_INTEGER_DTYPES:
-        raise TypeError(
-            f"expected signed integer train step indices from scheduler component "
-            f"{primary_name!r}, received dtype {indices.dtype}"
-        )
-    for name in group.names:
-        if name == primary_name:
-            continue
-        other = group[name].train_timesteps
-        if not isinstance(other, torch.Tensor) or not torch.equal(
-            other.to(indices.device), indices
-        ):
-            received = other.tolist() if isinstance(other, torch.Tensor) else other
-            raise ValueError(
-                "expected aligned scheduler-group train step indices; component "
-                f"{name!r} received {received} but primary {primary_name!r} declares "
-                f"{indices.tolist()}"
-            )
-    return indices
-
-
-def _build_training_component_times(
-    adapter: Any,
-    primary_timesteps: torch.Tensor,
-    *,
-    batch: Optional[StackedSampleBatch],
-) -> ComponentTimes:
-    if not isinstance(primary_timesteps, torch.Tensor):
-        raise TypeError(
-            "expected torch.Tensor primary_timesteps for build_training_component_times, "
-            f"received {type(primary_timesteps).__name__}: {primary_timesteps!r}"
-        )
-    if primary_timesteps.ndim != 1:
-        raise ValueError(
-            "expected primary_timesteps with one scheduler coordinate per sample, shape "
-            f"(B,), received {tuple(primary_timesteps.shape)}"
-        )
-    expected_names = ("latent",)
-    if adapter.trajectory_component_order != expected_names:
-        raise ValueError(
-            f"expected trajectory_component_order {expected_names} for the default "
-            f"build_training_component_times, received {adapter.trajectory_component_order}; "
-            "override the hook to map the primary coordinate per component"
-        )
-    sigma = flow_match_sigma(primary_timesteps)
-    return ComponentTimes(
-        timestep={"latent": primary_timesteps},
-        next_timestep={"latent": torch.zeros_like(primary_timesteps)},
-        sigma={"latent": sigma},
-        next_sigma={"latent": torch.zeros_like(sigma)},
-    )
-
-
-def _project_velocity_to_clean_state(
-    adapter: Any,
-    state: LatentState,
-    times: ComponentTimes,
-    velocity: LatentState,
-) -> LatentState:
-    """Project a noised state to clean data using the adapter's velocity direction."""
-    expected_names = adapter.trajectory_component_order
-    for argument, value in (("state", state), ("velocity", velocity)):
-        if not isinstance(value, LatentState):
-            raise TypeError(
-                f"expected LatentState for project_velocity_to_clean_state {argument}, "
-                f"received {type(value).__name__}"
-            )
-        if value.component_names != expected_names:
-            raise ValueError(
-                f"expected project_velocity_to_clean_state {argument} component order "
-                f"{expected_names}, received {value.component_names}"
-            )
-    if times.sigma is None or tuple(times.sigma) != expected_names:
-        received = None if times.sigma is None else tuple(times.sigma)
-        raise ValueError(
-            f"expected sigma component order {expected_names} for "
-            f"project_velocity_to_clean_state, received {received}"
-        )
-    direction = adapter.flow_velocity_direction
-    if direction not in ("noise", "data"):
-        raise ValueError(
-            "expected flow_velocity_direction to be 'noise' or 'data' for "
-            f"project_velocity_to_clean_state, received {direction!r}"
-        )
-
-    projected: Dict[str, torch.Tensor] = {}
-    for name in expected_names:
-        component_state = state.components[name]
-        component_velocity = velocity.components[name]
-        if component_state.ndim < 2:
-            raise ValueError(
-                f"expected project_velocity_to_clean_state component {name!r} to be batched "
-                f"with shape (B, ...), received {tuple(component_state.shape)}"
-            )
-        if component_velocity.shape != component_state.shape:
-            raise ValueError(
-                f"expected velocity component {name!r} to match state shape "
-                f"{tuple(component_state.shape)}, received {tuple(component_velocity.shape)}"
-            )
-        if component_velocity.device != component_state.device:
-            raise ValueError(
-                f"expected velocity component {name!r} on state device "
-                f"{component_state.device}, received {component_velocity.device}"
-            )
-        sigma = times.sigma[name]
-        batch_size = component_state.shape[0]
-        if sigma.ndim > 1 or sigma.numel() not in (1, batch_size):
-            raise ValueError(
-                f"expected sigma for component {name!r} to hold one value per sample with "
-                f"shape ({batch_size},), received {tuple(sigma.shape)}"
-            )
-        compute_dtype = torch.promote_types(component_state.dtype, component_velocity.dtype)
-        compute_state = component_state.to(dtype=compute_dtype)
-        compute_velocity = component_velocity.to(dtype=compute_dtype)
-        sigma = to_broadcast_tensor(sigma, compute_state)
-        sign = -1.0 if direction == "noise" else 1.0
-        projected[name] = compute_state + sign * sigma * compute_velocity
-    return LatentState(projected, active_masks=state.active_masks)
-
-
-def _apply_forward_process_noise(
-    adapter: Any,
-    clean_state: LatentState,
-    times: ComponentTimes,
-    noise: LatentState,
-) -> NoisedState:
-    expected_names = adapter.trajectory_component_order
-    for argument, state in (("clean_state", clean_state), ("noise", noise)):
-        if not isinstance(state, LatentState):
-            raise TypeError(
-                f"expected LatentState for apply_forward_process_noise {argument}, "
-                f"received {type(state).__name__}"
-            )
-        if state.component_names != expected_names:
-            raise ValueError(
-                f"expected apply_forward_process_noise {argument} component order "
-                f"{expected_names}, received {state.component_names}"
-            )
-    if times.sigma is None or tuple(times.sigma) != expected_names:
-        received = None if times.sigma is None else tuple(times.sigma)
-        raise ValueError(
-            f"expected sigma component order {expected_names} for "
-            f"apply_forward_process_noise, received {received}"
-        )
-    primary_name = expected_names[0]
-    primary_clean = clean_state.components[primary_name]
-    if primary_clean.ndim < 2:
-        raise ValueError(
-            f"expected apply_forward_process_noise clean_state component {primary_name!r} to be "
-            f"a batched tensor with a leading batch dimension and shape (B, ...), received shape "
-            f"{tuple(primary_clean.shape)}"
-        )
-    batch_size = primary_clean.shape[0]
-    noised: Dict[str, torch.Tensor] = {}
-    target_velocity: Dict[str, torch.Tensor] = {}
-    for name in expected_names:
-        clean_latents = clean_state.components[name]
-        component_noise = noise.components[name]
-        if clean_latents.ndim < 2 or clean_latents.shape[0] != batch_size:
-            raise ValueError(
-                f"expected batched clean_state component {name!r} with shape (B, ...) and "
-                f"batch size {batch_size}, received {tuple(clean_latents.shape)}"
-            )
-        if component_noise.shape != clean_latents.shape:
-            raise ValueError(
-                f"expected noise component {name!r} to match the clean shape "
-                f"{tuple(clean_latents.shape)}, received {tuple(component_noise.shape)}"
-            )
-        if (
-            component_noise.dtype != clean_latents.dtype
-            or component_noise.device != clean_latents.device
-        ):
-            raise ValueError(
-                f"expected noise component {name!r} to match the clean dtype/device "
-                f"({clean_latents.dtype}, {clean_latents.device}), received "
-                f"({component_noise.dtype}, {component_noise.device})"
-            )
-        sigma = times.sigma[name]
-        if sigma.ndim > 1 or sigma.numel() not in (1, batch_size):
-            raise ValueError(
-                f"expected sigma for component {name!r} to hold one value per sample with "
-                f"shape ({batch_size},), received {tuple(sigma.shape)}"
-            )
-        sigma = to_broadcast_tensor(sigma, clean_latents)
-        component_noised = (1 - sigma) * clean_latents + sigma * component_noise
-        component_target = component_noise - clean_latents
-        if clean_state.active_masks is not None:
-            # The draw above already consumed the full-shape RNG stream; masking only
-            # decides which elements move, so inactive conditioning stays clean and
-            # contributes no training signal.
-            mask = _expand_active_mask(
-                clean_state.active_masks[name],
-                clean_latents,
-                component=name,
-                operation="apply_forward_process_noise",
-            )
-            component_noised = torch.where(mask, component_noised, clean_latents)
-            component_target = torch.where(
-                mask, component_target, torch.zeros_like(component_target)
-            )
-        noised[name] = component_noised
-        target_velocity[name] = component_target
-    return NoisedState(
-        state=LatentState(noised, active_masks=clean_state.active_masks),
-        target_velocity=LatentState(target_velocity, active_masks=clean_state.active_masks),
-        noise=noise,
-    )
-
-
-def _add_forward_process_noise(
-    adapter: Any,
-    clean_state: LatentState,
-    times: ComponentTimes,
-    *,
-    generator: Optional[torch.Generator],
-) -> NoisedState:
-    expected_names = ("latent",)
-    if adapter.trajectory_component_order != expected_names:
-        raise ValueError(
-            f"expected trajectory_component_order {expected_names} for the default "
-            f"add_forward_process_noise, received {adapter.trajectory_component_order}; "
-            "override the hook to draw noise in component order"
-        )
-    if clean_state.component_names != expected_names:
-        raise ValueError(
-            f"expected exactly components {expected_names} for default noising, "
-            f"received {clean_state.component_names}"
-        )
-    if times.sigma is None or tuple(times.sigma) != expected_names:
-        received = None if times.sigma is None else tuple(times.sigma)
-        raise ValueError(
-            f"expected sigma component order {expected_names} before the default noise draw, "
-            f"received {received}"
-        )
-    clean_latents = clean_state.components["latent"]
-    # The single draw happens here, in component order, so the RNG stream stays
-    # reproducible; the application hook below consumes no randomness.
-    noise = randn_tensor(
-        clean_latents.shape,
-        generator=generator,
-        device=clean_latents.device,
-        dtype=clean_latents.dtype,
-    )
-    return adapter.apply_forward_process_noise(
-        clean_state,
-        times,
-        LatentState({"latent": noise}),
-    )
-
-
-def _build_forward_state_kwargs(
-    adapter: Any,
-    batch: StackedSampleBatch,
-    kwargs: Mapping[str, Any],
-) -> Dict[str, Any]:
-    """Resolve the forward arguments every ``forward_state`` override may pass on.
-
-    Owns the whole argument boundary: the state, times, return fields and noise
-    level belong to the bridge, and trajectory storage plus trainer metadata are
-    read from the batch rather than forwarded. Batch conditioning is resolved
-    first and explicit kwargs are layered on top; trainers drop the keys the batch
-    already carries before calling, which reproduces the legacy precedence of
-    ``forward(**training_args, **batch)``.
-    """
-    collisions = tuple(name for name in _STATE_OWNED_FORWARD_KEYS if name in kwargs)
-    if collisions:
-        raise ValueError(
-            f"explicit forward_state kwargs collide with state-owned arguments {collisions}"
-        )
-    owned = tuple(sorted(name for name in kwargs if name in _BRIDGE_OWNED_BATCH_KEYS))
-    if owned:
-        raise ValueError(
-            f"explicit forward_state kwargs collide with trainer-owned arguments {owned}; the "
-            f"{type(adapter).__name__} bridge reads trajectory storage and trainer metadata from "
-            "the batch and never forwards them to the model"
-        )
-    forward_kwargs = {
-        key: value
-        for key, value in batch.items()
-        if key not in _BRIDGE_OWNED_BATCH_KEYS and key not in _STATE_OWNED_FORWARD_KEYS
-    }
-    forward_kwargs.update(kwargs)
-    return forward_kwargs
-
-
-def _default_forward_state(
-    adapter: Any,
-    *,
-    batch: StackedSampleBatch,
-    state: LatentState,
-    times: ComponentTimes,
-    next_state: Optional[LatentState],
-    compute_log_prob: bool,
-    return_fields: Tuple[str, ...],
-    noise_level: Optional[float],
-    forward_kwargs: Mapping[str, Any],
-) -> MultiModalStepOutput:
-    expected_names = ("latent",)
-    received = {
-        "state": state.component_names,
-        "timestep": tuple(times.timestep),
-        "next_timestep": tuple(times.next_timestep),
-        "next_state": None if next_state is None else next_state.component_names,
-    }
-    if (
-        state.component_names != expected_names
-        or tuple(times.timestep) != expected_names
-        or tuple(times.next_timestep) != expected_names
-        or (next_state is not None and next_state.component_names != expected_names)
-    ):
-        raise ValueError(
-            f"expected exactly component order {expected_names} for default forward_state, "
-            f"received {received}"
-        )
-    output = adapter.forward(
-        t=times.timestep["latent"],
-        t_next=times.next_timestep["latent"],
-        latents=state.components["latent"],
-        next_latents=None if next_state is None else next_state.components["latent"],
-        compute_log_prob=compute_log_prob,
-        return_kwargs=return_fields,
-        noise_level=noise_level,
-        **filter_kwargs(adapter.forward, **forward_kwargs),
-    )
-    if not isinstance(output, SDESchedulerOutput):
-        raise TypeError(
-            "expected adapter.forward to return SDESchedulerOutput in forward_state, "
-            f"received {type(output).__name__}"
-        )
-
-    def wrap(value: Optional[torch.Tensor]) -> Optional[LatentState]:
-        return None if value is None else LatentState({"latent": value})
-
-    def wrap_statistic(value: Optional[torch.Tensor]) -> Optional[Dict[str, torch.Tensor]]:
-        return None if value is None else {"latent": value}
-
-    return MultiModalStepOutput(
-        next_state=wrap(output.next_latents),
-        next_state_mean=wrap(output.next_latents_mean),
-        std_dev_t=wrap_statistic(output.std_dev_t),
-        dt=wrap_statistic(output.dt),
-        log_prob=output.log_prob,
-        component_log_probs=wrap_statistic(output.log_prob),
-        velocity=wrap(output.velocity),
-    )
-
-
-def _resolve_component_latent_axes(
-    adapter: Any,
-    component: str,
-    latents: torch.Tensor,
-) -> LatentAxes:
-    if component != "latent":
-        raise ValueError(
-            "expected component 'latent' for default latent-axis resolution, "
-            f"received {component!r}"
-        )
-    return adapter.resolve_latent_axes(latents)
-
-
-def _validate_reduction_inputs(
-    adapter: Any,
-    values: Mapping[str, torch.Tensor],
-    state: Optional[LatentState],
-) -> int:
-    """Validate the shared reduction input contract and return the batch size."""
-    if not isinstance(values, Mapping) or not values:
-        raise ValueError(
-            f"expected non-empty Mapping[str, torch.Tensor] for values, received "
-            f"{type(values).__name__}"
-        )
-    expected_names = adapter.trajectory_component_order
-    if tuple(values) != expected_names:
-        raise ValueError(
-            f"expected values keys/order to match trajectory_component_order "
-            f"{expected_names}, received {tuple(values)}"
-        )
-    first = values[expected_names[0]]
-    if not isinstance(first, torch.Tensor):
-        raise TypeError(
-            f"expected batched torch.Tensor for values[{expected_names[0]!r}], "
-            f"received {type(first).__name__}"
-        )
-    if first.ndim < 1:
-        raise ValueError(
-            f"expected values[{expected_names[0]!r}] to be a batched torch.Tensor with a "
-            f"leading batch dimension, received shape {tuple(first.shape)}"
-        )
-    batch_size = first.shape[0]
-    if state is not None:
-        if not isinstance(state, LatentState):
-            raise TypeError(
-                f"expected LatentState or None for state, received {type(state).__name__}"
-            )
-        if state.component_names != expected_names:
-            raise ValueError(
-                f"expected state component order {expected_names}, "
-                f"received {state.component_names}"
-            )
-        for name, component_state in state.components.items():
-            if component_state.ndim < 1 or component_state.shape[0] != batch_size:
-                raise ValueError(
-                    f"expected state[{name!r}] to use the values batch size {batch_size}, "
-                    f"received shape {tuple(component_state.shape)}"
-                )
-    return batch_size
-
-
-def _validate_reduced_component_values(
-    adapter: Any,
-    reduced: Mapping[str, torch.Tensor],
-    batch_size: int,
-) -> Dict[str, torch.Tensor]:
-    """Validate a per-component reduction result before any trainer consumes it."""
-    identifier = f"reduce_component_latent_values on {type(adapter).__name__}"
-    if not isinstance(reduced, Mapping):
-        raise TypeError(
-            f"expected {identifier} to return Mapping[str, torch.Tensor], "
-            f"received {type(reduced).__name__}"
-        )
-    expected_names = adapter.trajectory_component_order
-    if tuple(reduced) != expected_names:
-        raise ValueError(
-            f"expected {identifier} to return component order {expected_names}, "
-            f"received {tuple(reduced)}"
-        )
-    expected_device: Optional[torch.device] = None
-    expected_dtype: Optional[torch.dtype] = None
-    for name in expected_names:
-        component_values = reduced[name]
-        if not isinstance(component_values, torch.Tensor):
-            raise TypeError(
-                f"expected {identifier} result [{name!r}] to be a torch.Tensor, "
-                f"received {type(component_values).__name__}"
-            )
-        if component_values.shape != (batch_size,):
-            raise ValueError(
-                f"expected {identifier} result [{name!r}] shape {(batch_size,)}, "
-                f"received {tuple(component_values.shape)}"
-            )
-        if expected_device is None:
-            expected_device = component_values.device
-            expected_dtype = component_values.dtype
-        elif component_values.device != expected_device or component_values.dtype != expected_dtype:
-            raise ValueError(
-                f"expected {identifier} result [{name!r}] to match component "
-                f"{expected_names[0]!r} ({expected_device}, {expected_dtype}), received "
-                f"({component_values.device}, {component_values.dtype})"
-            )
-    return dict(reduced)
-
-
-def _validate_reduced_latent_values(
-    adapter: Any,
-    reduced: torch.Tensor,
-    batch_size: int,
-) -> torch.Tensor:
-    """Validate a global reduction result before any trainer consumes it."""
-    identifier = f"reduce_latent_values on {type(adapter).__name__}"
-    if not isinstance(reduced, torch.Tensor):
-        raise TypeError(
-            f"expected {identifier} to return a torch.Tensor, received {type(reduced).__name__}"
-        )
-    if reduced.shape != (batch_size,):
-        raise ValueError(
-            f"expected {identifier} to return one scalar per sample with shape "
-            f"{(batch_size,)}, received {tuple(reduced.shape)}"
-        )
-    return reduced
-
-
-def _default_reduce_component_latent_values(
-    adapter: Any,
-    values: Mapping[str, torch.Tensor],
-    *,
-    state: Optional[LatentState],
-) -> Dict[str, torch.Tensor]:
-    active_masks = None if state is None else state.active_masks
-    reduced: Dict[str, torch.Tensor] = {}
-    for name in adapter.trajectory_component_order:
-        component_values = values[name]
-        if not isinstance(component_values, torch.Tensor) or component_values.ndim < 1:
-            raise ValueError(
-                f"expected values[{name!r}] to be a batched torch.Tensor with a leading "
-                f"batch dimension, received shape "
-                f"{tuple(getattr(component_values, 'shape', ()))}"
-            )
-        batch_size = component_values.shape[0]
-        if active_masks is None:
-            reduced[name] = component_values.reshape(batch_size, -1).mean(dim=1)
-            continue
-        mask = _expand_active_mask(
-            active_masks[name],
-            component_values,
-            component=name,
-            operation="reduce_component_latent_values",
-        )
-        counts = _active_element_counts(mask, component=name)
-        active_sum = (
-            torch.where(mask, component_values, torch.zeros_like(component_values))
-            .reshape(batch_size, -1)
-            .sum(dim=1)
-        )
-        reduced[name] = active_sum / counts.to(active_sum.dtype)
-    return reduced
-
-
-def _default_reduce_latent_values(
-    adapter: Any,
-    values: Mapping[str, torch.Tensor],
-    *,
-    active_numel: Optional[Mapping[str, int]],
-    state: Optional[LatentState],
-) -> torch.Tensor:
-    expected_names = adapter.trajectory_component_order
-    if active_numel is not None:
-        if not isinstance(active_numel, Mapping):
-            raise TypeError(
-                f"expected Mapping[str, int] or None for active_numel, "
-                f"received {type(active_numel).__name__}"
-            )
-        unknown_active_names = tuple(name for name in active_numel if name not in expected_names)
-        if unknown_active_names:
-            raise ValueError(
-                f"active_numel received unknown keys {unknown_active_names}; expected a subset "
-                f"of trajectory_component_order {expected_names}"
-            )
-
-    first = values[expected_names[0]]
-    if not isinstance(first, torch.Tensor) or first.ndim < 1:
-        raise TypeError(
-            f"expected batched torch.Tensor for values[{expected_names[0]!r}], "
-            f"received {type(first).__name__}"
-        )
-    batch_size = first.shape[0]
-    expected_device = first.device
-    expected_dtype = first.dtype
-    if len(expected_names) == 1 and first.ndim == 1:
-        override = None if active_numel is None else active_numel.get(expected_names[0])
-        if override is not None:
-            if not isinstance(override, int) or isinstance(override, bool) or override <= 0:
-                raise ValueError(
-                    f"expected active_numel[{expected_names[0]!r}] to be a positive int, "
-                    f"received {override!r}"
-                )
-        return first
-
-    active_masks = None if state is None else state.active_masks
-    weighted_sum: Optional[torch.Tensor] = None
-    total_weight: Union[int, torch.Tensor] = 0
-    for name in expected_names:
-        component_values = values[name]
-        if not isinstance(component_values, torch.Tensor) or component_values.ndim < 1:
-            raise TypeError(
-                f"expected batched torch.Tensor for values[{name!r}], "
-                f"received {type(component_values).__name__}"
-            )
-        if component_values.shape[0] != batch_size:
-            raise ValueError(
-                f"expected batch size {batch_size} for values[{name!r}], received "
-                f"shape {tuple(component_values.shape)}"
-            )
-        if component_values.device != expected_device or component_values.dtype != expected_dtype:
-            raise ValueError(
-                f"expected compatible dtype/device from component "
-                f"{expected_names[0]!r} ({expected_device}, {expected_dtype}), received "
-                f"{name!r} ({component_values.device}, {component_values.dtype})"
-            )
-        override = None if active_numel is None else active_numel.get(name)
-        if override is not None:
-            if not isinstance(override, int) or isinstance(override, bool) or override <= 0:
-                raise ValueError(
-                    f"expected active_numel[{name!r}] to be a positive int, "
-                    f"received {override!r}"
-                )
-            if component_values.ndim != 1:
-                raise ValueError(
-                    f"expected already-reduced values[{name!r}] shape (B,) when "
-                    f"active_numel is provided, received {tuple(component_values.shape)}"
-                )
-            component_sum = component_values * override
-            component_weight: Union[int, torch.Tensor] = override
-        elif active_masks is not None:
-            mask = _expand_active_mask(
-                active_masks[name],
-                component_values,
-                component=name,
-                operation="reduce_latent_values",
-            )
-            component_weight = _active_element_counts(mask, component=name)
-            component_sum = (
-                torch.where(mask, component_values, torch.zeros_like(component_values))
-                .reshape(batch_size, -1)
-                .sum(dim=1)
-            )
-        else:
-            flattened = component_values.reshape(batch_size, -1)
-            component_sum = flattened.sum(dim=1)
-            component_weight = flattened.shape[1]
-        if not isinstance(component_weight, torch.Tensor) and component_weight <= 0:
-            raise ValueError(
-                f"expected positive element weight for component {name!r}, "
-                f"received {component_weight}"
-            )
-        weighted_sum = component_sum if weighted_sum is None else weighted_sum + component_sum
-        total_weight = total_weight + component_weight
-    if isinstance(total_weight, torch.Tensor):
-        total_weight = total_weight.to(weighted_sum.dtype)
-    elif total_weight <= 0:
-        raise ValueError(f"expected positive total latent weight, received {total_weight}")
-    return weighted_sum / total_weight
